fix: cleanup duplicate type definitions across packages
- Remove Skywalker.Ddd.Domain.Entities.EntityNotFoundException; unify on Skywalker.Exceptions.EntityNotFoundException
- Remove unused Skywalker.Logging.IHasLogLevel + HasLogLevelExtensions
- Move EntityChangedEvent<T> family from Skywalker.EventBus.Local to Skywalker.Ddd.Domain.Events
- Rename Skywalker.Ddd.AspNetCore.Mvc.Models.Abstractions.ValidationError to ValidationErrorDto
- Fix EntityAlreadyExistedException xml-doc cref typo

Closes #194
