+++ src/Skywalker.Ddd.AspNetCore/Skywalker/Ddd/AspNetCore/Mvc/Models/Abstractions/Error.cs
@@ -29,7 +29,7 @@ public class Error
     /// <summary>
     /// Validation errors if exists.
     /// </summary>
-    public ValidationError[]? ValidationErrors { get; set; }
+    public ValidationErrorDto[]? ValidationErrors { get; set; }
 
 #pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider adding the 'required' modifier or declaring as nullable.
     [JsonConstructor]
