+++ src/Skywalker.Ddd.AspNetCore/Skywalker/Ddd/AspNetCore/Mvc/Models/Abstractions/ValidationErrorDto.cs
@@ -1,4 +1,4 @@
-﻿// Licensed to the Gordon under one or more agreements.
+// Licensed to the Gordon under one or more agreements.
 // Gordon licenses this file to you under the MIT license.
 
 using System.Text.Json.Serialization;
@@ -6,10 +6,10 @@ using System.Text.Json.Serialization;
 namespace Skywalker.Ddd.AspNetCore.Mvc.Models.Abstractions;
 
 /// <summary>
-/// Used to store information about a validation error.
+/// HTTP response DTO that carries a single validation error to the client.
 /// </summary>
 [Serializable]
-public class ValidationError
+public class ValidationErrorDto
 {
 
     /// <summary>
@@ -24,15 +24,15 @@ public class ValidationError
 
 #pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider adding the 'required' modifier or declaring as nullable.
     [JsonConstructor]
-    private ValidationError() { }
+    private ValidationErrorDto() { }
 #pragma warning restore CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider adding the 'required' modifier or declaring as nullable.
 
     /// <summary>
-    /// Creates a new instance of <see cref="ValidationError"/>.
+    /// Creates a new instance of <see cref="ValidationErrorDto"/>.
     /// </summary>
     /// <param name="member">Related invalid member</param>
     /// <param name="message">Validation error message</param>
-    public ValidationError(string member, string message)
+    public ValidationErrorDto(string member, string message)
     {
         Member = member;
         Message = message;
