--- src/Skywalker.Extensions.Universal/Skywalker/Logging/IHasLogLevel.cs
@@ -1,14 +1,0 @@
-using Microsoft.Extensions.Logging;
-
-namespace Skywalker.Logging;
-
-/// <summary>
-/// Interface to define a <see cref="LogLevel"/> property (see <see cref="LogLevel"/>).
-/// </summary>
-public interface IHasLogLevel
-{
-    /// <summary>
-    /// Log severity.
-    /// </summary>
-    LogLevel LogLevel { get; set; }
-}
