+++ tests/Skywalker.EventBus.Tests/DomainEventTests.cs
@@ -73,24 +73,6 @@ public class DomainEventTests
         Assert.Empty(entity.DomainEvents);
     }
 
-    [Fact]
-    public void EntityChangedEvent_SetsProperties()
-    {
-        // Arrange
-        var entity = new TestEntity();
-
-        // Act
-        var createdEvent = new EntityCreatedEvent<TestEntity>(entity);
-        var updatedEvent = new EntityUpdatedEvent<TestEntity>(entity);
-        var deletedEvent = new EntityDeletedEvent<TestEntity>(entity);
-
-        // Assert
-        Assert.Same(entity, createdEvent.Entity);
-        Assert.Equal(EntityChangeType.Created, createdEvent.ChangeType);
-        Assert.Equal(EntityChangeType.Updated, updatedEvent.ChangeType);
-        Assert.Equal(EntityChangeType.Deleted, deletedEvent.ChangeType);
-    }
-
     public class TestDomainEvent : DomainEventBase
     {
         public string Value { get; set; } = string.Empty;
